server/Records/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["RavinduJS <110709934+RavinduJS@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.8"}
boto3 = "^1.36.18"
pydantic = "^2.10.6"
strawberry-graphql = {extras = ["fastapi"], version = "^0.260.2"}
pymongo = "^4.11.1"

[tool.poetry.scripts]
dev = 'app.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
